Move all in-memory functionalit to impl
