+++ src/jevent/mixin/EventStore.ts
@@ -1,6 +1,6 @@
-import { getOrCreateObjectStore, getOrCreateStringStore, StoreProvider } from '../../store/Store'
 import { BaseConstructor } from '../JEvent'
 import { EventStore } from '../EventStore'
+import { StoreProvider, getOrCreateStringStore, getOrCreateObjectStore } from '../../store/impl/InMemoryStore'
 
 const WithEventStore = <TBase extends BaseConstructor>(Base: TBase, storeProvider: StoreProvider) => {
   return class extends Base implements EventStore {
